finalisation site statique

+++ src/Entity/Drink.php
@@ -20,7 +20,7 @@ class Drink
     private ?string $size = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $img = null;
@@ -60,12 +60,12 @@ class Drink
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(?int $price): self
+    public function setPrice(?float $price): self
     {
         $this->price = $price;
